ai-service/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
whisper
numpy
python-dotenv